Magicbane is a quarterstaff.
From EvilHack/FIQHack. I think this change works well for the spellcasters, especially priests, wizards, and healers. But it's also a fact that athames are not really as important as they were in 3.4.3, making Magicbane not as potent as it could be. With the better damage of a quarterstaff and the spellcasting boost, it should make Magicbane much more appealing.