Quest artifacts cannot be wished for.
This is ported from dNetHack with the implementation derived from EvilHack.

This is partly to make players strategize around their role's quest artifact more, but also to make adding new roles a bit less power-unbalancing. I found from adding several roles in the past that each new role adds to the pool of uber-powerful quest artifacts. With many roles added, a player has a buffet of artifacts to choose from, making wishes even more over-powered than they already are. I would like to be able to add roles in the future without this concern, so I'm addressing it now.