Crowning gifts are only granted when crowned, never for #offer.
This is something I always thought was a bit strange with the crowning gifts, that you could receive one before actually crowned. It always made the actual crowning a bit less exciting. So we've restricted these gifts to the main event:
- Priests don't get Mjollnir from #offering
- Chaotics don't get Stormbringer from #offering
- Neutrals don't get Vorpal Blade from #offering
- Lawfuls don't get Excalibur from #offering

This does a couple of things: It further pushes Excalibur out of easy reach for most lawfuls, it makes Mjollnir more special for priests (it is really good now so it should feel that way), but it also puts more pressure on the player to get a crowning from the altar before it cracks. This might be a tricky puzzle but I think they'll figure it out ;)