Revert "Restricted or unskilled throwing has a chance of weapon slipping (from UnNetHack)."
This reverts commit 2747b9fc1f1d73248ef8565ec5e362129d2ada5c.

Reverting this because it's another change that punishes unskilled weapons. This turns out to be pretty rough for early game and it's just not fun to deal with. It also makes the slipping mechanic less intuitive, does a slipping projectile mean that it's cursed or you are unskilled?? These two factors are enough to revert it.