Peaceful monsters won't cast make invisible on themselves.
With the previous setting, a peaceful monster would only cast Disappear on themselves if the player had See Invisible. However, since it's quite difficult to get permanent see invisible now, the player will likely timeout their intrinsic see invisible or take off the item granting it.

A funny example that made me realize this was the titan on Medusa's level. I was wearing a ring of see invisible when I entered the level, and the titan was luckily peaceful, but it made itself invisible at some point. I reentered the level later, wearing a different ring and ran straight into it!