Invisibility protects from gaze attacks.
As long as a monster doesn't see invisible, you have a 1 in 11 chance of avoiding a gaze attack. I noticed that blinking eyes had 100% accuracy even when the player was blind, which just didn't make sense. I used these odds because its the same number that monmove uses when monsters are trying to detect you. A buff to invisibility feels nice because it's not possible to gain permanent intrinsic invisibility anymore. It also makes the ring and cloak of invisibility more appealing.

Future followups could also handle displacement and being in dark areas when facing monsters without infravision.