Darkness offers protection from gaze attacks.
This is another area where gazing monsters seemed to have an unfair advantage. Now, if you are on a dark square and the gazer doesn't have infravision, they won't be able to target you. However, if they are in melee range, they will be able to reasonably spot you and target you again.

If they do have infravision, you must be a form that is infravisible for them to target you.

I think this change makes it a bit more fair, but it also makes carrying a light source expose you to more gaze attacks. So traveling in the dark isn't as bad as it used to be.