New artifact weapon: Load Brand.
Introducing a really heavy two-handed sword that packs a wallop. This is kind of a callback to the Bat from Hell, but a version of it that is a bit more balanced.

It weighs just as much as a load stone - so 500. But in addition to the standard two-handed sword damage it deals 1d20 with 1d5 to-hit bonus. It also grants steadfastness and MC1 protection. I added a new item, the heavy sword, that is basically a copy of the two-handed sword, but it's made of mineral and of course the weight is changed. We could have manually tweaked the item weight and reused the two-handed sword, but this leaves room for further tweaks down the line.

This weapon could be a blessing or a curse for players, but it definitely can clear out baddies fast so we'll see!