Reflection only provides partial protection from floating eye gazes.
This is a fairly soft nerf, but I believe it lines up with the behavior of other gaze attacks. Reflection has no affect on any other gaze (except Medusa's gaze) so the fact that it fully protects from the floating eye feels inconsistent.

Now, when a player gets the passive paralysis gaze from a floating eye and possesses reflection, they will be paralyzed for 1d6 turns. This is still massively better than without reflection, but the player should still exercise some caution when around floating eyes.