Nerf: Reflecting Medusa's gaze.
From EvilHack. Reflecting her gaze back from more than 3 squares away is ineffective. If you do reflect her gaze, it only works 2% of the time (1 in 50). This renders reflection as more of a protective measure. In EvilHack this was 1 in 12, but I think she would be even better at protecting herself from this attack, and it will be more interesting to kill her in other ways. There are some other buffs to Medusa, but we'll wait on those for now.