Further reduce the damage gems/stones do to thick skinned monsters.
My original nerf turned out not to be quite sufficient. I was easily able to defeat the Chromatic Dragon as a caveman by boulder forting and slinging flint. It actually didn't take that many flint stones to my surprise. The damage is handled more directly now so that a max of 1 damage is possible.