Decrease the protection that can be bought from priests or granted by your god.
From K-Mod. The probability of being granted protection decreases with each point.

No more 2-4 points of AC on the first gift. The calculation now limits the maximum divine protection to 10 points. When donating or receiving a divine protection gift, the chance of failure is calculated as follows:

Current 	Chance
protection      of fail
0 		0%
1 		1%
2 		4%
3 		9%
4 		16%
5 		25%
6 		36%
7 		49%
8 		64%
9 		81%
10 		100%