Strength gain nerfs. Giant corpses have less of a chance of conferring strength gain (from 50% to 17%).
When the player does gain strength (from spinach/giant corpse/royal jelly/etc), the gains are halved when they hit the greater gain. To clarify, when the player is under 18 STR, they had a 25% chance to gain 1d6 STR, but now it is 1d3. At the next tier (18 to 18(85) STR), they could gain 1d10 but now it's throttled to 1d5.

The giant nerf is adapted from SLASH'EM, where it was nerfed to 25%.