Monsters benefit from the racial armor bonuses.
This is to mirror the behavior for the player. This might significantly boost early game dwarf and gnome AC values so I also removed the berserking AC boost. This way berserking dwarves won't just be unstoppable juggarnauts. This should result in tougher opposition across the board: orcs, elves, gnomes, and dwarves should be noticeably more difficult, so this will require some playtesting to see if it's too much.