Armor weight adjustments.
Using some inspiration from K-Mods adjustments, I reduced most of the armor weights by about 50. With the worn armor weight discount of 25%, this should help to free up some inventory space for players.

On the other hand, I increased the weights of mithril and dragon scales so the 'regular' armors are a bit more competitive.