Opened up a bunch of race/role combos.
- Elvish healer
- Dwarven knight
- Dwarven priest
- Orcish priest
- Elvish rogue
- Gnomish rogue
- Elvish tourist
- Gnomish tourist
- Dwarven barbarian

I had a decision on whether to convert small shields to their racial equivalent (ie: to Elven shield if elf), but I passed on this so that priests and valkyries always get their normal small shields. For valkyries this means the +3 shields won't be replaced by a dwarven roundshield with an additional -1AC for dwarves. For priests, they can be reassured they can cast spells with less resistance since the racial shields all stubbornly block spellcasting