Increase chances of getting a golem from polypiling (from SporkHack).
This is adapted from Spork, but we use a different formula.

Spork used: if (!rn2(Luck/5 + 2))

However, this felt a bit excessive, giving the player at best (13/5 + 2) or 1 in 4 odds of a golem.

We are using: if (!rn2(Luck + 7))

So that the player has 1 in (13 + 7) (or 1 in 20) of a golem when Luck is maxed out. Because luck times out this feels fair but it still pushes golems much more often than Vanilla does, further making polypiling a strategy of desperation rather than staple.