Throne wishes have been removed. To compensate, full recursive identify of items can occur if the player has enough luck.
The throne wish case will only grant luck now. The case for identifying items "you are granted an insight", now also has a chance to grant the player an epiphany. The chain identify is from FIQHack and can be quite useful if you have a large bag with lots of unknown items.