Cavemen can receive slashing/piercing weapons gifts again.
This was a restriction that was meant to align with the priest's inability to unrestrict edged weapon skills. But, it makes sense with the weird way cavepeople gods work that they can get these gifts, just not the skills. Cavepeople will get strange gifts, but it may work out sometimes if they can overcome the to-hit penalties. It also means they can get Dragonbane, which can be pretty helpful for their quest.