HP gains from healing potions are subject to nurse dancing limits.
This is a topic that was tossed around in the #nethack channel and I think there was much validity to it. The max reasonable HP with this implementation is a little above 500HP at level 30. This should be quite sufficient for most characters.

I can see some followup adjustments to this:

1) Partial HP gains are not handled. Currently the player has a check based on their level. They either get the full max HP gain from the potion or nothing. Perhaps we could handle fractions later.
2) Should the max HP gains from potions be more or less than nurse dancing? I could see an argument for either direction, but for now, having the same limit across the board seems the most logical.