Adult dragon buffs (from SLASH'EM).
All dragons are now level 18, difficulty 25, have at least -4AC, get a 4d8 bite and 3d4 claw attacks, and are frequency 2. This is to complement the baby dragon changes. Part of the challenge in SLASH'EM was that adult dragons (and therefore dragon scales) were more difficult to come by. Baby dragons are plentiful and can potentially be tamed and grown up into adults, but with the large gap in levels the process is long and difficult. It's possible they might not be guaranteed in Ludios or the Castle anymore either, but the increased frequency means they should show up in Gehennom much more often.