Demonic bribes are much more expensive.
Many variants do a spin on this concept. I'm keeping it simple and just charging the shit out of the player so that gold has some use for the late game.

Expect to see bribe demands of $4000-$10000. If your alignment matches the demon's, you'll get a bit of a discount.

I used EvilHack and xNetHack as guides for coding this.