Shorten floating eye paralysis, wisdom limits duration (from DynaHack).
From commit 46c218038:

'Floating eyes taking a well-equipped full health character to instant
death was absurd, so make it less likely and easier for certain
characters to mitigate.

The passive paralysis of floating eyes is now (HD+1)d35 instead of
(HD+1)d70, where hit dice is 2 but can vary depending on when they're
encountered.

Passive paralysis no longer has a 1 in 4 chance of being maxed out to
127 turns instead of its intended duration.  Instead of wisdom
preventing this, it now sets the maximum turns to 10 * (20 - wisdom),
down to a minimum of 10 turns.'