Dragons, nagas, and golems don't balk at approaching as much.
This was actually caused by a recent 3.7 change to make monsters with ranged attacks retreat more often. Unfortunately it does a couple of annoying things. It makes sacrificing nagas and dragons on altars almost impossible because they simply won't die on the altar. This is because they alswayswill retreat when under 1/3rd of their health. It also makes killing the new faster dragons much more difficult because they also stay just out of reach and pepper the player with breath blasts (the black dragons are really nasty at this!)

To fix this, I made a couple of tweaks. First, I reduced the threshold for retreating to 25% of health instead of 33%. This should make most monsters with ranged attacks fight in melee a bit longer. Second, I added exceptions for dragons, nagas, and golems to fight in melee more often. These monsters don't feel so much like retreating monsters to me. Dragons are naturally aggressive, nagas are... weird, and golems should be mostly mindless and robotic. They can still balk and retreat, but much less.