Explicit summon nasties message when it occurs as part of wizard harassment.
After the player kills the wizard, a crowd of nasties can occasionally be silently amassed around the player. This change gives the player a little warning message, similar to the message when a spellcaster or Rodney summons nasties.

As a helpful hint, this lets you put 'Monsters appear from nowhere!' in your MSGTYPE stops or alerts so you don't inadvertantly move after getting surrounded.