Sceptre of Might buff.
Gave the Sceptre a flat +3 bonus, it felt pretty puny during my failed caveman run, so let's see if this carries it through the endgame.

Some issues (as a neutral caveman) were having to punt at air and earth elementals for 10 to even 15 hits. This was awful. As a lawful they are targeted much more severely, but we should at least be able to get through them with a little less effort.