Replaced levitation boots with flying boots.
That's right! RIP levitation boots.

Many changes have proven that flying is more important than ever. Post-invocation earthquakes, lava in gehennom, and getting smashed and hurtled by monsters are all lethal threats. I think converting the levitation boots to an item that is actually usable is a great player-friendly move.

I had also thought of importing the cloak of flight from SpliceHack, but I think this is a more efficient use of items.