Adjust altar #offer odds in favor of the player
We kind of want players to crack the altars instead of endlessly sacrificing for nothing. Here are the adjusted odds for the first 10 gifts. They are quite a big higher, so players should have to think about things like getting gifts vs getting crowned, or risking destroying the mines altar (and losing a source of divine protection.)

Gifts
Granted 	Old odds	New odds
0		10.00%		12.50%
1		8.33%		10.00%
2		5.56%		8.33%
3		3.57%		7.14%
4		2.38%		6.25%
5		1.67%		5.56%
6		1.22%		5.00%
7		0.93%		4.55%
8		0.72%		4.17%
9		0.58%		3.85%
10		0.48%		3.57%