Bane artifacts can instakill their associated monster types.
Ported in from EvilHack and UnNetHack. I saw a real need for this type of adjustment because it seemed absurd that a bane like Ogresmasher would require over 5 hits when unenchanted to kill an ogre. I saw a couple of ways that different variants dealt with this. FIQhack made it so the banes all dealt d20. This adjustment was appealing, but I still liked the idea of instakill and we already had disrupter with a d30 bonus, and giving other banes a d20 felt incompatible with that. My main interest was making the banes viable without needing to enchant them too heavily. Because resources have been constrained so much, we shouldn't require the player to waste too many enchant weapon scrolls on a bane to make it do what it should inherently do.

The main adjustment from EvilHack is that the banes have a 1 in 5 chance of instakilling, both vs monsters and the player. EvilHack has the odds at 10% vs monsters, and 5% vs the player. Since we don't allow two-weaponing, I think we can use the straight 20% odds from UnNetHack. It's definitely possible the player might get on the wrong side of one of these, but unlikely since we don't have demonic or giant races (yet).

I also fixed some is_race macros that were out of sync with the recent MH flag updates.