Baby dragons get a claw attack.
I changed their bite to 2d3 and added a 1d6 physical claw attack. This makes their overall damage a little stronger (from an average of 7 damage to 7.5). If the player is immune to their element, they should have a little backup. If they prove to be too strong for the early game, we might have to tweak the attacks or their difficulty slightly.