Confused cursed destroy armor prompts for armor to fix.
This syncs up the behavior between enchant armor and destroy armor. If we don't do this, players might get confused and use confused destroy armor expecting the behavior of confused enchant armor. The main condition for this to work is that the player must have previously identified the scroll of destroy armor. Simply calling the scroll is not enough either.

This is also a bit less lenient than the confused enchant armor, you only get one try to pick the armor you want, otherwise the scroll is wasted.