Introducing Mayhem - chaotic artifact stomping boots.
I was mulling over new artifacts to import and noticed that chaotics don't seem to get much for armor in variants. So... at first the idea of some "Dark Soles" came to mind, but of course that was a bit cheesy, so we have Mayhem instead. A major part of the inspiration also was directed to creating an anti-Whisperfeet.

When worn, Mayhem conveys conflict and warning vs undead. In addition, the random appearance of the boots might also have a significant benefit. The conflict is not controllable (like the Sceptre of Might), but this is by design and a minor downside. These boots also offer chaotics a way to navigate the hoards of undead in Gehennom, something that usually only lawful artifacts offer.