Ported Deluder (from SLASH'EM).
This is also a decent artifact cloak. I made one tweak and removed the Luck bonus and replaced it with a level of protection. This kind of brings it back to the MC2 players enjoyed in SLASH'EM. For some reason, most of the artifact armors in SLASH'EM also granted luck, so I'm backing away from that.