Ported Mirrorbright (from SLASH'EM).
This is a neutral shield of reflection that also grants hallucination resistance when worn.

In SLASH'EM, this was a fairly mediocre item, but we have added two major buffs that make it stand out:
1) It doesn't impede spellcasting. This change was taken from slashem-up.
2) It acts as a light source. This is a new change and a play on the name since it has 'bright' in the title.

In SLASH'EM it was also the first sac gift for healers, but in the interest of gift variety we have refrained from that guarantee.