Ported Blackshroud (from SlashTHEM).
In SlashTHEM, this was a neutral cloak of protection that granted luck, drain resistance, and warning.

We have tweaked it a bit, now it is a cloak of invisiblity that grants drain resistance and warning. The cloak of protection provides MC3, which is quite valuable, but it also provides most of the protection that drain res does. Invisiblity is a more valuable intrinsic in this variant and it fits better with warning. Luck was removed because it was one property too much and it's a running theme of SLASH artifacts to always provide luck.