Ported Snakeskin (from SlashTHEM).
This is the last of the neutral armors I'll import from SLASH variants, but this one I couldn't resist.

In SlashTHEM this is a neutral robe that confers hallucination resistance and acid resistance when worn. It also granted protection, but I left that out since the robe already has MC2.

This is the only artifact that grants acid resistance, which is a quite a valuable property with all the new item destruction and acid blasts spellcasters can throw at you. Will it be as legendary as EvilHack's Dragonbane???  Only time will tell ;)