Half physical/spell damage nerf.
Instead of 1/2, they reduce the damage by 1/4.

I simply adjusted the calculation wherever I could for Half_physical_dmg or Half_spell_dmg. There were a couple of niche spots that were inpractical to adjust but probably nothing the player will ever notice.

Both of these effects are fairly overpowered, so I'm going to try an experiment to see if only reducing a 1/4 of the damage results in more interesting games. Half physical damage basically doubles your AC which either makes a mediocre AC great, or a great AC unstoppable. With this approach, I'm hoping to make the power jump not as steep.

In addition, most artifacts granting 1/2 damage reduction even grant it slotless!

On the bright side, I believe this nerf makes it viable to add half physical damage to the Tsurugi of Muramasa and the Longbow of Diana. Both of these are two-handed weapons and historically a little underpowered.