Caveman quest updates (from SlashTHEM).
I noticed that THEM had a bunch of interesting changes. Notably, a lot of water is added and bugbears are replaced with more jungle animals.

The core quest monsters are still bugbear(h) and hill giant(H), but the initial wave is a bunch of jungle monsters. Giant eels, electric eels, and pythons are also added to hide in the water.