Sokoban levels are cold and icy.
As you ascend the Soko puzzles, the tower gets colder and more floor tiles are replaced with ice. The top 3 levels are tagged as 'cold'. Right now I'm not sure if this actually does anything, but I predict the 3.7 devteam will be doing more with this. My initial assumption was that potions dropped on cold levels could be destroyed (like scrolls in hot hell levels), but currently this does not seem to be the case.