New monster: like-like.
Inspired by the NES version of Legend of Zelda, this is a cake-like monster that can engulf you and eat your shields and cloaks.

If you don't have a shield or cloak to eat, then it either deals rotting damage during the engulf or as part of it's touch attack. These are bad news because they can even eat your erodeproofed items. Luckily, silver shields are immune to this attack. Artifact cloaks and shields are still susceptible to this attack but as part of the destroy armor coding they will resist 90% of the time.

I put them into the pudding (P) category because that felt the closest to whatever this thing is. Like green slimes, these also only appear in hell - they feel a bit too nasty for the main dungeon.

As part of their attack, I updated destroy_arm to take a new boolean dictating whether it should print with the function is called.

Also - as a callback to the NES, they can only move cardinally like grid bugs.