New monster: shadow ogre (from SLASH'EM).
One more ogre for the lineup - and this one might start to make Ogresmasher look a little more attractive.

These are ported with some changes.
- They are higher level at 13, they were level 10 in SLASH'EM. Their difficulty is also rated higher.
- They can generate randomly in the main dungeon. In SLASH'EM they only were coded to spawn in Grund's Stronghold.
- They now get a crushing bearhug
- They always spawn invisible. I thought about some complicated mechanics with darkness, but as usual the simpler option appeals to me more here.
- Instead of being traitor, they get flanking and berserk.
- They resist draining.

This monster serves a couple of purposes. It puts more strain on the player's newfound weakness in seeing invisible. Most characters will not have intrinsic see inv, so will be relying on rings or warning. Secondly, these also drain life which is not an easy thing to defend against. Those two combined makes them a nasty surprise for the late-mid game, so they should be popping up somewhere before the castle.