New item: anti-magic shield.
This is basically re-purposing the hide shield from SpliceHack into a shield of magic resistance. I used the same stats and tiles so credit goes to Splice for that. It is a leather shield that grants 2AC and magic resistance.

I was debating porting in an amulet of magic resistance, but I think the shield works better. Plus, since we have like-likes now, they will be happy to eat these up.

Currently I don't have any ideas for monsters that could spawn with this, but perhaps ogre mages...