New item: gauntlets of force.
These are kind of an evolution from the boxing gloves in SpliceHack and the gauntlets of force in SporkHack.

- These are leather gloves that provide an increase damage (similar to rings of increase damage).
- When worn and fighting bare-handed, you can force-fight and instantly smash any iron bars or boulders in your way.
- You will also have a 1 in 40 chance of inflicting a staggering blow on your opponents.
- When forcing open boxes and chests, these will force them open in 1 turn.
- You are guaranteed to open any unlocked door.

Their main strength is in their increase damage potential. Since they are leather, they are probably more valuable to spellcasters. I noticed that some variants let the player get a +x damage bonus for the enchantment on their gloves. I like this concept but it lets monks start off with a +2 bonus because of their gloves which results in them steamrolling the early game. I also think that allowing this bonus to occur on all gloves is a bit overpowered, especially on gauntlets of dexterity, which should be a defensive item. The bonus lets those gloves become a powerhouse of offense and defense.

So that is the main inspiration for these gloves - to allow a damage bonus, but only on a special pair of gloves. I think the universal damage bonus is much more appealing than the more limited bare-handed bonus that the above mentioned gloves would get. This makes them practical for more than just monks.

Being able to break through iron bars is also pretty nice with a lot of post 3.7 changes. Also - it might let the player skip a bunch of Soko levels...