New monster spell: Evil Eye (from dNetHack).
This is a mage-spell that drains the hero's luck by 1 if it connects.

I had thought of porting in barghests or evil eyes (which have a luck draining gaze), but this is a bit more efficient and lets any mage-spellcaster cast it.

Since it calls the monster gaze attack the player can protect themselves by using standard gaze protection measures: invisibility, displacement, being in darkness, and the Eyes of the Overworld offer varying migitation.