New monster spell: entomb (from xNetHack).
This mage-spell is used by a monster when they are low on health or fleeing. It is primarily for escape and surrounds the player with boulders. I made a few tweaks to this from xnh. This spell isn't limited to Dispater, any mage-spellcaster can use it when desperate. There was also the chance to create solid walls to block the player in, but I think the boulders will be enough so I removed that feature.

I tried to limit the frequency of the spell a bit, but this will still require playtesting to tune in so the player isn't just spammed with boulders.