Potions can shatter when dropped on cold floor.
This is the counterpart change to the 3.7 feature where potions can heat up and break from hot ground.

Since the upper floors of Soko are cold, the player will have to be a bit more careful. The entry level is still not cold, so a decent place for a stash.