Elven gear weighs less (from EvilHack).
The weights of all elven items have been reduced by about 1/3.

This is a cool feature I noticed in Evil in one commit, but the change there was to fix issues with object materials and elven items. We don't have object materials, but I like the idea that elves would be more efficient, especially since elven gear is now more limited to elves and humans.