more conversion cracking
Just to push this concept even further, when a player converts an altar, there is a coin flip to see if it cracks or not. If minions are summoned in retaliation, the altar is always cracked. However, the minion summoning can usually be mitigated by having good alignment.

I keep pushing this because I want to see what players can get by with when forced to 'play the hand they are dealt'. I predict that it will still be easy to get at least 4-5 artifacts, giving the player a fair shake at something endgame worthy. Even if they don't get altar-lucky, there is always the option to wish for a fancy artifact...