Reword blocking messages.
paxxed pointed out that the blocking message for boots was a bit weird:

'Your pair of stomping boots deflects the athol's attack.'

so I reworked it so it would instead say

'You block the athol's attack with your pair of stomping boots.'

It's a bit more of an active than a passive stance. After all, WE are doing the blocking, not our armor.