Red dragon tweaks.
Turns out this was the deadliest thing in the game with the buffs I gave it. And it turns out that I don't *need* to buff it's attacks when it already has berserk, because all it's attacks effectively double after it berserks LOL.

I brought it's attacks and AC back down to match the other dragons. With it's speed it should still have quite high DPS when berserking.