Intrinsic teleportitis is temporary.
Normally this is obtained by eating leprechaun, nymph, or tengu corpses and grants permanent teleportitis. This can be quite annoying if it wasn't intentional, especially for newer players who don't know all the ins-and-outs of corpse intrinsics.

Now, eating a tele-corpse grants 500-750 turns of teleportitis. This seems like enough to get out of the troubling condition easy, but also enough to make use of for travelling.

It also makes eating tengu corpses early much more appealing.