Almost all items are erodeable/destroyable.
Ported from EvilHack.

This makes it so that almost any item can be affected by erosion. It's actually kind of a reversion to the way 3.4.3 worked, in that all items could become eroded back then, but it didn't really matter since it was merely cosmetic. They removed that feature in 3.6 to the dismay of some player, but I think there is a lot of interesting possibilities with more vulnerable items and making erosion really matter.

One function that is really missing is a generic erode_items function. There is a burn_armor method to handle armor because there is complexity in the 3 layers and it needs it's own special handling. But, for other items and damage like acid and rotting there isn't a good alternative. For this commit, I kind of kludged my way into the maybe_destroy_item method and added handling for also eroding items. This way, we don't need to add a bunch of extra calls and a simple destroy_items call will do both, instantly destroying vulnerable items and eroding other items. For now, I think this will work, but I would be open to better options in the future.

Many new avenues of corrosion and rotting were opened up. Previously, we opened up more acid damage, but now in addition the rest of the inventory is susceptible.

A few ways were added to deal with erosion:
- Previous: Forges can be used to remove erosion from metallic items, and even have a one-time chance to erodeproof an item.
- Previous: Potions of restore ability can remove all the erosion from an item
- New: Confused scrolls of enchant weapon can be used to erodeproof any wieldable item.

Followup commits will make eroded tools less effective, finetune the probabilities of erosion, and tweak other eroded item behavior.