Food items rotted from effects also trigger the rotten food effects.
This was a bit weird, since the rotted erosion usually goes on oeroded2 and the standard eating doesn't really check that since it assumed it was a separate set of effects.

I just added checks to eat.c to look at oeroded2. This could probably also be done when the erosion is applied, but this seemed like the simplest approach for now.