Unskilled unicorn horn/healer buff.
If you have unskilled or restricted skill in unicorn horns, this reduces your success rate with them by one notch. This requires roles that can't achieve skill in unihorn to invest some enchantment, or encourages roles that can to spend some skill points.

If you are a healer with a +0 unicorn horn, it acts like a +2 horn for 50% success.