Fix: dipping worn non-metallic items into forges.
The previous behavior was that non-metal items would get instantly burned up in the forge lava, but strangely you wouldn't take any damage. I updated the control flow so that dipping any worn item of any material deals lava damage, and updated the message for when it hits the lava to be a little less verbose and clunky.