Black dragons get a terrifying roar.
From xNetHack and Splice. In SpliceHack all the dragons can emit a frightful roar, scaring the player. In xnethack, they just roar but it doesn't scare the player.

I really like this mechanic, but I want it to be special to one dragon. The black dragon is one of the most dangerous with its disintegration breath so we're giving it the roar so it can also disorient players.

Wielding Dragonbane offers full protection from this attack.

I was able to reuse the existing nazgul shrieking code because they both work almost identically. I made a new function because blending the two attacks ended up just being awkward, and we can fine-tune the behavior of each more if they are separate.