While the player is carrying the Amulet of Yendor, monsters flood from the upstairs.
This is adapted from EvilHack, but without the monclock changes.

Once the player has the amulet in possession, there is a 1 in 10 chance each turn of a monster spawning to block the <. This does directly call summon nasty, so it could be a little over-the-top. In Evil, the monsters are spawned using makemon but the probability should be the same.

I guess the player could just drop the amulet to stop monsters from generating, but eventually they will be pressed forward anyway.