Lessen confusing messages for confuse monster effects.
I'm not sure why we needed a message everytime we hit something when we have the confuse monster hands - because it makes it feel like it wears off every time. I changed it so you only get 1 warning when there are 2 turns left, and a final warning on the last charge.