Player race elves get see invisible at level 19.
It confused me that player elves didn't get see invisible when all the monster elves get it, so let's fix that. I went with level 19 since some roles get an intrinsic at level 20 and it's a little less message spam to give one at a time.

Yea it's high, but I can't justify giving it to them earlier without throwing something else at dwarves, orcs, or gnomes, or the poor humans. It gives elves something to go for.