Ensure healer quest gets adult dragons.
The baby dragon changes seem to shift the balance of D on the quest to babies. I noticed in SLASH'EM this was never addressed, so they get a bit of a cakewalk. We match the random D with guaranteed adult dragons so there should be a good mix of baby and adults.