Disintegration rays can vaporize boulders.
I begin with the assumption that boulders are big enough to block our sight. From there, logically, we can assume that if they are tall enough to block sight, they are also tall enough to be hit by a ray and therefore would vaporize.

This is probably only really nasty in regards to the Chromatic Dragon if a player is relying on a boulder fort, but it also obviously affects defense against black dragons.