Summon demon lords/princes when you possess the Amulet of Yendor.
This takes the place of the demon lairs. Now when you are carrying the amulet, part of the wizard's harassment is now summoning these lords/princes. Some of the demons are still bribable, so I lowered the bribe money a little since it's more unpredictable and carrying that much cash in the open is not easy.