Potion HP gains use max level for ceiling.
We are still using the nurse dancing ceiling, but by setting the ceiling to use the max level of 30, we allow low level characters to safely gain max HP.

I noticed this at level 1 when I tried to quaff a potion of extra healing to gain 2 HP with a tourist. I had 14HP and nothing happened. Quite dissapointing for a squishy character like that.

With the new calculation, the average ceiling is 480HP. Getting above that is definitely possile but an uphill battle.