Ranseurs can disarm enemies (suggested by disperse).
Really cool idea for the oft-neglected polearms. If used in melee or pounding, ranseurs can not only knock enemies weapons to the ground, but also pry off their shields.

Currently the item is merely knocked onto the ground in the same square the monster occupies. Ideally, it would be knocked a square to the side so they can't just pick it up immediately.