Spears at expert skill can skewer through enemies.
This is inspired and adapted from a YANI writeen by aosdict and shadowrider (https://nethack-yanis.github.io/yanis/4259.html).

When your role is at expert with spears and not two-weaponing, they can get a double hit on the monster behind the primary target. This works similarly to how Cleaver cleaves, and in fact, I mostly duplicated the code for cleaving and simplified it to one target.

The check for hitting peacefuls also follows Cleavers direction - if the spear is cursed however, all bets are off and you will skewer through to it anyway.

Currently only three roles can reach expert in spears: valkyries, cavemen/cavewomen, and rangers. I think it's worth revisiting this skill so a couple other roles could enjoy this mechanic.