New artifact: Serenity.
This is a lawful silver spear. Silver spears are pretty awesome, but unfortunately quite rare. Having an artifact silver spear gives the player a higher chance of actually using one. I always liked the Holy Spear of Light from SLASH'EM - but it's function is mostly duplicated by Sunsword and/or Disrupter. So we're starting from scratch and redesigning a silver spear with it's own origin and function.

It's now a fact that berserking monsters are a potent threat. Even in the endgame with a strong AC, the player is going to have to deal with some monsters that normally aren't very damaging, but once they berserk, are kind of insane (red dragons are a prime example). The very root of this artifact is to act as a counter to berserkers, as wielding this artifact will prevent berserkers from berserking.

Serenity's base material as silver is always great for extra damage, but we'll grant it a standard +d5 to-hit and +d10 damage to make it competitive.

Its secondard ability is simple but useful - it grants extrinsic fire resistance when wielded. It's fairly easy to gain intrinsic fire resistance, but that doesn't protect your inventory from fire damage. This will.

It's final ability is also something new. I wanted something like spell damage reduction, but instead, we're going to let this act as a pre-emptive counterspell so wielding it both absorbs curses and also prevents many monsters from casting spells in the first place. If this is cursed, it loses it's ability to counter spells.